[fix] 리뷰 Response 값 수정, 재료 쿼리문 수정 요청 완료

+++ src/main/java/store/teabliss/review/dto/ReviewCreateDto.java
@@ -14,7 +14,7 @@ public class ReviewCreateDto {
     private String contents;
 
     @Schema(description = "별점", example = "5")
-    private Long likes;
+    private int likes;
 
     @Schema(description = "완제품 차 고유 번호", example = "1")
     private Long teaId;
